pyproject.toml
[project]
name = "canvas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.0",
    "pillow>=12.1.0",
]
